+++ packages/SUI.BACKEND/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.19.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.18.0...v5.19.0) (2019-10-14)
+
+**Note:** Version bump only for package @smsoft/sui-backend
+
+
+
+
+
+# [5.18.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.17.1...v5.18.0) (2019-10-14)
+
+
+### Features
+
+* **backendtable:** sortingEnabled && Service columns ([1e5ee75](https://github.com/mbuyakov/SUI.CORE/commit/1e5ee75))
+
+
+
+
+
 ## [5.17.1](https://github.com/mbuyakov/SUI.CORE/compare/v5.17.0...v5.17.1) (2019-10-09)
 
 **Note:** Version bump only for package @smsoft/sui-backend