+++ packages/SUI.PROMISED/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.19.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.18.0...v5.19.0) (2019-10-14)
+
+**Note:** Version bump only for package @smsoft/sui-promised
+
+
+
+
+
+# [5.18.0](https://github.com/mbuyakov/SUI.CORE/compare/v5.17.1...v5.18.0) (2019-10-14)
+
+**Note:** Version bump only for package @smsoft/sui-promised
+
+
+
+
+
 ## [5.14.10](https://github.com/mbuyakov/SUI.CORE/compare/v5.14.9...v5.14.10) (2019-10-02)
 
 **Note:** Version bump only for package @smsoft/sui-promised